implement icon cache expiry, cleanup code and upgrade deps
